+++ skforecast/direct/_forecaster_direct.py
@@ -1490,7 +1490,8 @@ class ForecasterDirect(ForecasterBase):
         self,
         steps: int | list[int] | None = None,
         last_window: pd.Series | pd.DataFrame | None = None,
-        exog: pd.Series | pd.DataFrame | None = None
+        exog: pd.Series | pd.DataFrame | None = None,
+        check_inputs: bool = True
     ) -> pd.DataFrame:
         """
         Create the predictors needed to predict `steps` ahead.
@@ -1514,6 +1515,10 @@ class ForecasterDirect(ForecasterBase):
             right after training data.
         exog : pandas Series, pandas DataFrame, default None
             Exogenous variable/s included as predictor/s.
+        check_inputs : bool, default True
+            If `True`, the input is checked for possible warnings and errors 
+            with the `check_predict_input` function. This argument is created 
+            for internal use and is not recommended to be changed.
 
         Returns
         -------
@@ -1531,7 +1536,8 @@ class ForecasterDirect(ForecasterBase):
         ) = self._create_predict_inputs(
                 steps        = steps,
                 last_window  = last_window,
-                exog         = exog
+                exog         = exog,
+                check_inputs = check_inputs
             )
 
         X_predict = pd.DataFrame(